refactor: phase 1 - split raid data access into model/view/domain layers
