+++ app/components/features/raids/RaidCard.tsx
@@ -23,7 +23,7 @@ import {
 } from "~/locales/ko";
 import { bossImageUrl } from "~/models/assets";
 import type { RaidType, Terrain } from "~/models/content.d";
-import type { Difficulty } from "~/models/raid";
+import type { Difficulty } from "~/domain/raid-score";
 import { sanitizeClassName } from "~/prophandlers";
 
 type RaidCardProps = {
