+++ app/models/raid.ts
@@ -5,7 +5,6 @@ import { type UtcIsoString, compareInstantAsc, isInstantAfter, nowUtcIso, toUtcI
 import { cacheKey, cacheQuery, fetchSourceCached } from "~/lib/cache";
 import { getTimelineContentDatesByContentUids } from "./timeline-content";
 
-export type Difficulty = "normal" | "hard" | "very_hard" | "hardcore" | "extreme" | "insane" | "torment" | "lunatic";
 const ALL_RAID_SCHEDULES_CACHE_KEY = cacheKey("source", "raid-schedule", 1, cacheQuery({ region: "gl" }));
 
 type RawRaidSchedule = NonNullable<Awaited<ReturnType<typeof runRaidScheduleDetailQuery>>>;
@@ -167,6 +166,9 @@ export function getAllRaidSchedules(env: Env, forceRefresh = false) {
   );
 }
 
+export type RaidSchedule = NonNullable<Awaited<ReturnType<typeof getRaidSchedule>>>;
+export type RaidScheduleListItem = Awaited<ReturnType<typeof getAllRaidSchedules>>[number];
+
 export async function getUpcomingRaidSchedules(env: Env, forceRefresh = false) {
   const now = nowUtcIso();
   const schedules = await getAllRaidSchedules(env, forceRefresh);
@@ -175,129 +177,60 @@ export async function getUpcomingRaidSchedules(env: Env, forceRefresh = false) {
     .sort((a, b) => compareInstantAsc(a.startAt ?? now, b.startAt ?? now));
 }
 
-const TOTAL_ASSAULT_BOSS_UIDS = [
-  "binah",
-  "chesed",
-  "hod",
-  "shirokuro",
-  "perorodzilla",
-  "goz",
-  "hieronymus",
-  "kaiten",
-  "gregorius",
-  "hovercraft",
-  "kurokage",
-  "geburah",
-  "yesod",
-  "drumbarka",
-] as const;
-
-export type Boss = (typeof TOTAL_ASSAULT_BOSS_UIDS)[number];
-
-export const ALL_TOTAL_ASSUALT_BOSS: Boss[] = [...TOTAL_ASSAULT_BOSS_UIDS];
-
-const LEGACY_BOSS_UIDS: Partial<Record<string, Boss>> = {
-  perorozilla: "perorodzilla",
-  "kaiten-fx-mk0": "kaiten",
-  "myouki-kurokage": "kurokage",
-};
-
-export function normalizeBossUid(boss: string): Boss | null {
-  if ((TOTAL_ASSAULT_BOSS_UIDS as readonly string[]).includes(boss)) {
-    return boss as Boss;
+export async function findRaidScheduleSummaryByTypeAndSeason(
+  env: Env,
+  raidType: string,
+  seasonIndex: number | string,
+  forceRefresh = false,
+) {
+  const parsedSeasonIndex = typeof seasonIndex === "number" ? seasonIndex : Number.parseInt(String(seasonIndex), 10);
+  if (Number.isNaN(parsedSeasonIndex)) {
+    return null;
   }
-  return LEGACY_BOSS_UIDS[boss] ?? null;
+
+  const schedules = await getAllRaidSchedules(env, forceRefresh);
+  return schedules.find((schedule) => schedule.raidType === raidType && schedule.seasonIndex === parsedSeasonIndex) ?? null;
 }
 
-const RAID_TIME_SCORE_PER_SECOND = {
-  normal: 120,
-  hard: 240,
-  very_hard: 480,
-  hardcore: 960,
-  extreme: 1440,
-  insane: 1920,
-  torment: 2400,
-  lunatic: 2880,
-};
+export async function getRaidScheduleByTypeAndSeason(
+  env: Env,
+  raidType: string,
+  seasonIndex: number | string,
+  forceRefresh = false,
+) {
+  const summary = await findRaidScheduleSummaryByTypeAndSeason(env, raidType, seasonIndex, forceRefresh);
+  if (!summary) {
+    return null;
+  }
 
-const RAID_DIFFICULTY_SCORE = {
-  normal: 250000,
-  hard: 500000,
-  very_hard: 1000000,
-  hardcore: 2000000,
-  extreme: 4000000,
-  insane: 6800000,
-  torment: 12200000,
-  lunatic: 17710000,
-};
-
-const RAID_HP_SCORE = {
-  normal: { 180: 229000, 240: 277000, 270: 304700 },
-  hard: { 180: 458000, 240: 554000, 270: 609400 },
-  very_hard: { 180: 916000, 240: 1108000, 270: 1218800 },
-  hardcore: { 180: 1832000, 240: 2216000, 270: 2437600 },
-  extreme: { 180: 5392000, 240: 6160000, 270: 6578880 },
-  insane: { 180: 12449600, 240: 14216000, 270: 14941016 },
-  torment: { 180: 18876000, 240: 19508000, 270: 20302000 },
-  lunatic: { 180: 25525000, 240: 26315000, 270: 26954000 },
-};
-
-function timeForBoss(boss: Boss): 180 | 240 | 270 {
-  if (boss === "yesod" || boss === "drumbarka") return 270;
-  if (boss === "binah" || boss === "kaiten") return 180;
-  return 240;
+  return getRaidSchedule(env, summary.uid, forceRefresh);
 }
 
-export function timeToScore(boss: Boss, difficulty: Difficulty, clearTimeMillisec: number): number {
-  if (clearTimeMillisec < 0 || clearTimeMillisec > 3600000) {
-    throw new Error("소요 시간은 0 ~ 60분 사이여야 해요");
-  }
-
-  const timePerBoss = timeForBoss(boss);
-  const hpScore = RAID_HP_SCORE[difficulty][timePerBoss];
-  if (hpScore === null) {
-    throw new Error("선택한 보스는 아직 해당 난이도가 개최되지 않았어요");
-  }
-
-  const difficultyScore = RAID_DIFFICULTY_SCORE[difficulty];
-  const timeScore = ((3600000 - clearTimeMillisec) / 1000) * RAID_TIME_SCORE_PER_SECOND[difficulty];
-  return Math.round(difficultyScore + hpScore + timeScore);
+export async function findRaidScheduleSummaryByTypeAndJpSeason(
+  env: Env,
+  raidType: string,
+  jpSeasonIndex: number,
+  forceRefresh = false,
+) {
+  const schedules = await getAllRaidSchedules(env, forceRefresh);
+  return (
+    schedules.find(
+      (schedule) => schedule.raidType === raidType && schedule.jpSchedule?.seasonIndex === jpSeasonIndex,
+    ) ?? null
+  );
 }
 
-export function scoreToDifficultyAndTime(
-  boss: Boss,
-  score: number,
-): { difficulty: Difficulty; clearTimeMillisec: number } {
-  const timePerBoss = timeForBoss(boss);
-  const difficulties: Difficulty[] = [
-    "normal",
-    "hard",
-    "very_hard",
-    "hardcore",
-    "extreme",
-    "insane",
-    "torment",
-    "lunatic",
-  ];
+export async function warmRaidCache(env: Env) {
+  const now = nowUtcIso();
+  const schedules = await getAllRaidSchedules(env, true);
 
-  // Try each difficulty from highest to lowest
-  for (let i = difficulties.length - 1; i >= 0; i--) {
-    const difficulty = difficulties[i];
-    const hpScore = RAID_HP_SCORE[difficulty][timePerBoss];
-    if (hpScore === null) {
-      continue;
-    }
+  await Promise.all(
+    schedules
+      .filter((schedule) => schedule.endAt && isInstantAfter(schedule.endAt, now))
+      .map((schedule) => getRaidSchedule(env, schedule.uid, true)),
+  );
 
-    const difficultyScore = RAID_DIFFICULTY_SCORE[difficulty];
-    const timeScore = score - difficultyScore - hpScore;
-
-    const clearTimeMillisec = Math.round(3600000 - (timeScore * 1000) / RAID_TIME_SCORE_PER_SECOND[difficulty]);
-    if (clearTimeMillisec >= 0 && clearTimeMillisec <= 3600000) {
-      return { difficulty, clearTimeMillisec };
-    }
-  }
-
-  throw new Error("유효하지 않은 점수에요");
+  return schedules;
 }
 
 // ============================================================
@@ -327,28 +260,4 @@ export async function applyTimelineDateFallback<
       endAt: s.endAt ?? dates?.endAt ?? null,
     };
   });
-}
-
-// ============================================================
-// URL param ↔ internal raidType conversion
-// ============================================================
-
-const PARAM_TO_RAID_TYPE: Record<string, string> = {
-  "total-assault": "total_assault",
-  "grand-assault": "elimination",
-};
-
-const RAID_TYPE_TO_PARAM: Record<string, string> = {
-  total_assault: "total-assault",
-  elimination: "grand-assault",
-};
-
-/** URL path param (e.g. "total-assault") → internal raidType (e.g. "total_assault") */
-export function raidTypeFromParam(param: string): string {
-  return PARAM_TO_RAID_TYPE[param] ?? param;
-}
-
-/** Internal raidType (e.g. "elimination") → URL path param (e.g. "grand-assault") */
-export function raidTypeToParam(raidType: string): string {
-  return RAID_TYPE_TO_PARAM[raidType] ?? raidType;
 }
