+++ app/repositories/index.ts
@@ -1,14 +1,9 @@
-export { RaidRepository } from "./raid";
 export {
   DEFAULT_VIDEO_SORT,
   isVideoSort,
   parseVideoSort,
   RAID_VIDEOS_PAGE_SIZE,
 } from "~/models/raid-videos";
-export type {
-  RaidSchedule,
-  RaidScheduleListItem,
-} from "./raid";
 export type {
   RaidVideoItem,
   RaidVideosData,
