+++ app/routes/raids.$raidType.$seasonIndex._index.tsx
@@ -6,10 +6,11 @@ import RaidClearLevels from "~/components/features/raids/RaidClearLevels";
 import RaidOftenUsedParties from "~/components/features/raids/RaidOftenUsedParties";
 import RaidStatisticsSlotCount from "~/components/features/raids/RaidStatisticsSlotCount";
 import { EmptyView, HorizontalScroll, LoadingSkeleton, Section } from "~/components/primitives";
+import { raidTypeToParam } from "~/domain/raid";
 import { fetchRaidOverview } from "~/lib/ranks/overview";
 import { type RaidStatistics, fetchRaidStatisticsByRaid } from "~/lib/ranks/stats";
 import type { RaidType } from "~/models/content.d";
-import { getRaidDefenseTypeSetKey, raidTypeToParam } from "~/models/raid";
+import { getRaidDefenseTypeSetKey } from "~/models/raid";
 import { getSameOccurrenceRaids } from "~/models/raid-group";
 import { getMaxTierAt } from "~/models/student";
 import { getAllStudentsMap } from "~/models/student";
