+++ app/routes/students.$id._components/StudentDifficultyUsageModel.ts
@@ -1,7 +1,7 @@
 import type { Defense } from "~/graphql/graphql";
 import type { StudentAnalysisResponse } from "~/lib/ranks/student-analysis";
 import type { RaidType, Terrain } from "~/models/content.d";
-import type { Difficulty } from "~/models/raid";
+import type { Difficulty } from "~/domain/raid-score";
 
 export type StudentAnalysisScopeMetadata = {
   bossName: string;
